+++ isa/src/main/java/rs/ftn/isa/model/Room.java
@@ -41,7 +41,7 @@ public class Room {
 	private int sprat;
 	//cijena za noc
 	@Column(name = "cijena", nullable = false)	
-	private int cijena;
+	private double cijena;
 
 	@Column(name = "kapacitet", nullable = false)	
 	private int kapacitet; //da ako ima,ne nema
@@ -73,7 +73,7 @@ public class Room {
 	public Room() {
 		super();
 	} 
-	public Room(Long id, String tip, double ocjena, int kreveti, int sprat, int cijena,String balkon) {
+	public Room(Long id, String tip, double ocjena, int kreveti, int sprat, double cijena,String balkon) {
 		super();
 		this.id = id;
 		this.tip = tip;
@@ -97,7 +97,7 @@ public class Room {
 		this.balkon = balkon;
 	}
 	
-	public Room( String tip, double ocjena, int kreveti, int sprat, int cijena,String balkon) {
+	public Room( String tip, double ocjena, int kreveti, int sprat, double cijena,String balkon) {
 		super();
 		this.tip = tip;
 		this.ocjena = ocjena;
@@ -107,7 +107,7 @@ public class Room {
 		this.balkon = balkon;
 	}
 	
-	public Room(String tip, double ocjena, int kreveti, int sprat, int cijena) {
+	public Room(String tip, double ocjena, int kreveti, int sprat, double cijena) {
 		super();
 		this.tip = tip;
 		this.ocjena = ocjena;
@@ -116,7 +116,7 @@ public class Room {
 		this.cijena = cijena;
 	}
 
-	public Room(String tip,int kapacitet,  int kreveti, int sprat, int cijena,String balkon) {
+	public Room(String tip,int kapacitet,  int kreveti, int sprat, double cijena,String balkon) {
 		super();
 		this.tip = tip;
 		this.ocjena = 0;
@@ -167,11 +167,11 @@ public class Room {
 		this.sprat = sprat;
 	}
 
-	public int getCijena() {
+	public double getCijena() {
 		return cijena;
 	}
 
-	public void setCijena(int cijena) {
+	public void setCijena(double cijena) {
 		this.cijena = cijena;
 	}
 	public Hotel getHotel() {
